INS-309: move jet coordinator interface to core/ledger
