+++ core/ledger.go
@@ -16,6 +16,16 @@
 
 package core
 
+type JetRole int
+
+const (
+	RoleVirtualExecutor  = JetRole(iota + 1) // Role responsible for current CPU operations
+	RoleVirtualValidator                     // Role responsible for past CPU operations
+	RoleLightExecutor                        // TODO: add docs
+	RoleLightValidator                       // TODO: add docs
+	RoleHeavyExecutor                        // TODO: add docs
+)
+
 // Ledger is the global ledger handler. Other system parts communicate with ledger through it.
 type Ledger interface {
 	// GetArtifactManager returns artifact manager to work with.
@@ -26,4 +36,12 @@ type Ledger interface {
 
 	// GetPulseManager returns pulse manager to work with.
 	GetPulseManager() PulseManager
+}
+
+type JetCoordinator interface {
+	// IsAuthorized checks for role on concrete pulse for the address
+	IsAuthorized(role JetRole, obj RecordRef, pulse PulseNumber, node RecordRef) bool
+
+	// TODO: add docs
+	QueryRole(role JetRole, obj RecordRef, pulse PulseNumber) []RecordRef
 }
